[Response Ops] [Rule Form] Remove V1 Rule Form Flyout (#209171)
## Summary

Closes #195211

Removes all old rule form flyout code, which should no longer be
referenced at all after https://github.com/elastic/kibana/pull/206685/

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

(cherry picked from commit fffc18cfc424c52a1d635c789927e4c479cbb5e7)

# Conflicts:
#	x-pack/examples/alerting_example/tsconfig.json
#	x-pack/platform/plugins/shared/triggers_actions_ui/.storybook/decorator.tsx
#	x-pack/platform/plugins/shared/triggers_actions_ui/common/experimental_features.ts
#	x-pack/platform/plugins/shared/triggers_actions_ui/public/application/rules_app.tsx
#	x-pack/platform/plugins/shared/triggers_actions_ui/public/application/sections/rules_list/components/rules_list.tsx
#	x-pack/platform/plugins/shared/triggers_actions_ui/public/common/get_experimental_features.test.tsx
