+++ x-pack/examples/alerting_example/public/alert_types/always_firing.tsx
@@ -18,9 +18,6 @@ import {
 import { i18n } from '@kbn/i18n';
 import { omit, pick } from 'lodash';
 import {
-  ActionGroupWithCondition,
-  AlertConditions,
-  AlertConditionsGroup,
   RuleTypeModel,
   RuleTypeParamsExpressionProps,
 } from '@kbn/triggers-actions-ui-plugin/public';
@@ -29,6 +26,7 @@ import {
   AlwaysFiringActionGroupIds,
   DEFAULT_INSTANCES_TO_GENERATE,
 } from '../../common/constants';
+import { RuleConditions, RuleConditionsGroup, ActionGroupWithCondition } from '../components';
 
 export function getAlertType(): RuleTypeModel {
   return {
@@ -101,7 +99,7 @@ export const AlwaysFiringExpression: React.FunctionComponent<
       <EuiSpacer size="m" />
       <EuiFlexGroup>
         <EuiFlexItem grow={true}>
-          <AlertConditions
+          <RuleConditions
             headline={'Set different thresholds for randomly generated T-Shirt sizes'}
             actionGroups={actionGroupsWithConditions}
             onInitializeConditionsFor={(actionGroup) => {
@@ -111,7 +109,7 @@ export const AlwaysFiringExpression: React.FunctionComponent<
               });
             }}
           >
-            <AlertConditionsGroup
+            <RuleConditionsGroup
               onResetConditionsFor={(actionGroup) => {
                 setRuleParams('thresholds', omit(thresholds, actionGroup.id));
               }}
@@ -124,8 +122,8 @@ export const AlwaysFiringExpression: React.FunctionComponent<
                   });
                 }}
               />
-            </AlertConditionsGroup>
-          </AlertConditions>
+            </RuleConditionsGroup>
+          </RuleConditions>
         </EuiFlexItem>
       </EuiFlexGroup>
       <EuiSpacer />
