+++ x-pack/platform/plugins/shared/triggers_actions_ui/.storybook/decorator.tsx
@@ -71,7 +71,6 @@ export const StorybookContextDecorator: FC<PropsWithChildren<StorybookContextDec
       ruleKqlBar: true,
       isMustacheAutocompleteOn: false,
       showMustacheAutocompleteSwitch: false,
-      isUsingRuleCreateFlyout: false,
     },
   });
