+++ x-pack/platform/plugins/shared/triggers_actions_ui/public/application/home.tsx
@@ -72,7 +72,6 @@ export const TriggersActionsUIHome: React.FunctionComponent<RouteComponentProps<
       'xl'
     )({
       showCreateRuleButtonInPrompt: true,
-      useNewRuleForm: true,
       setHeaderActions,
     });
   }, []);
