+++ x-pack/platform/plugins/shared/triggers_actions_ui/public/application/rules_app.tsx
@@ -54,14 +54,12 @@ import { KibanaContextProvider, useKibana } from '../common/lib/kibana';
 import { ConnectorProvider } from './context/connector_context';
 import { ALERTS_PAGE_ID, CONNECTORS_PLUGIN_ID } from '../common/constants';
 import { queryClient } from './query_client';
-import { getIsExperimentalFeatureEnabled } from '../common/get_experimental_features';
 
 const TriggersActionsUIHome = lazy(() => import('./home'));
 const RuleDetailsRoute = lazy(
   () => import('./sections/rule_details/components/rule_details_route')
 );
-const CreateRuleRoute = lazy(() => import('./sections/rule_form/rule_form_route'));
-const EditRuleRoute = lazy(() => import('./sections/rule_form/rule_form_route'));
+const RuleFormRoute = lazy(() => import('./sections/rule_form/rule_form_route'));
 
 export interface TriggersAndActionsUiServices extends CoreStart {
   actions: ActionsPublicPluginSetup;
@@ -125,25 +123,19 @@ export const AppWithoutRouter = ({ sectionsRegex }: { sectionsRegex: string }) =
     application: { navigateToApp },
   } = useKibana().services;
 
-  const isUsingRuleCreateFlyout = getIsExperimentalFeatureEnabled('isUsingRuleCreateFlyout');
-
   return (
     <ConnectorProvider value={{ services: { validateEmailAddresses } }}>
       <Routes>
-        {!isUsingRuleCreateFlyout && (
-          <Route
-            exact
-            path={createRuleRoute}
-            component={suspendedComponentWithProps(CreateRuleRoute, 'xl')}
-          />
-        )}
-        {!isUsingRuleCreateFlyout && (
-          <Route
-            exact
-            path={editRuleRoute}
-            component={suspendedComponentWithProps(EditRuleRoute, 'xl')}
-          />
-        )}
+        <Route
+          exact
+          path={createRuleRoute}
+          component={suspendedComponentWithProps(RuleFormRoute, 'xl')}
+        />
+        <Route
+          exact
+          path={editRuleRoute}
+          component={suspendedComponentWithProps(RuleFormRoute, 'xl')}
+        />
         <Route
           path={`/:section(${sectionsRegex})`}
           component={suspendedComponentWithProps(TriggersActionsUIHome, 'xl')}
