+++ x-pack/platform/plugins/shared/triggers_actions_ui/public/application/sections/rule_details/components/rule.tsx
@@ -201,7 +201,6 @@ export function RuleComponent({
           actionTypeRegistry,
           ruleTypeRegistry,
           hideEditButton: true,
-          useNewRuleForm: true,
           onEditRule: requestRefresh,
         })}
       </EuiFlexGroup>
