+++ x-pack/platform/plugins/shared/triggers_actions_ui/public/application/sections/rule_details/components/rule_details.tsx
@@ -6,7 +6,7 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import React, { useState, useEffect, useReducer, useMemo } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import {
   EuiPageHeader,
@@ -52,14 +52,12 @@ import {
 } from '../../common/components/with_bulk_rule_api_operations';
 import { RuleRouteWithApi } from './rule_route';
 import { ViewInApp } from './view_in_app';
-import { RuleEdit } from '../../rule_form';
 import { routeToRules } from '../../../constants';
 import {
   rulesErrorReasonTranslationsMapping,
   rulesWarningReasonTranslationsMapping,
 } from '../../rules_list/translations';
 import { useKibana } from '../../../../common/lib/kibana';
-import { getRuleReducer } from '../../rule_form/rule_reducer';
 import { loadAllActions as loadConnectors } from '../../../lib/action_connector_api';
 import { runRule } from '../../../lib/run_rule';
 import {
@@ -71,7 +69,6 @@ import {
 import { useBulkOperationToast } from '../../../hooks/use_bulk_operation_toast';
 import { RefreshToken } from './types';
 import { UntrackAlertsModal } from '../../common/components/untrack_alerts_modal';
-import { getIsExperimentalFeatureEnabled } from '../../../../common/get_experimental_features';
 
 export type RuleDetailsProps = {
   rule: Rule;
@@ -79,7 +76,6 @@ export type RuleDetailsProps = {
   actionTypes: ActionType[];
   requestRefresh: () => Promise<void>;
   refreshToken?: RefreshToken;
-  useNewRuleForm?: boolean;
 } & Pick<
   BulkOperationsComponentOpts,
   'bulkDisableRules' | 'bulkEnableRules' | 'bulkDeleteRules' | 'snoozeRule' | 'unsnoozeRule'
@@ -102,7 +98,6 @@ export const RuleDetails: React.FunctionComponent<RuleDetailsProps> = ({
   const {
     application: { capabilities, navigateToApp },
     ruleTypeRegistry,
-    actionTypeRegistry,
     setBreadcrumbs,
     chrome,
     http,
@@ -111,14 +106,6 @@ export const RuleDetails: React.FunctionComponent<RuleDetailsProps> = ({
     userProfile,
     notifications: { toasts },
   } = useKibana().services;
-
-  const isUsingRuleCreateFlyout = getIsExperimentalFeatureEnabled('isUsingRuleCreateFlyout');
-
-  const ruleReducer = useMemo(() => getRuleReducer(actionTypeRegistry), [actionTypeRegistry]);
-  const [{}, dispatch] = useReducer(ruleReducer, { rule });
-  const setInitialRule = (value: Rule) => {
-    dispatch({ command: { type: 'setRule' }, payload: { key: 'rule', value } });
-  };
 
   const [rulesToDelete, setRulesToDelete] = useState<string[]>([]);
   const [rulesToUpdateAPIKey, setRulesToUpdateAPIKey] = useState<string[]>([]);
@@ -178,7 +165,6 @@ export const RuleDetails: React.FunctionComponent<RuleDetailsProps> = ({
       ? !ruleTypeRegistry.get(rule.ruleTypeId).requiresAppContext
       : false);
 
-  const [editFlyoutVisible, setEditFlyoutVisibility] = useState<boolean>(false);
   const onRunRule = async (id: string) => {
     await runRule(http, toasts, id);
   };
@@ -240,10 +226,6 @@ export const RuleDetails: React.FunctionComponent<RuleDetailsProps> = ({
     hasEditButton,
   ]);
 
-  const setRule = async () => {
-    history.push(getRuleDetailsRoute(rule.id));
-  };
-
   const goToRulesList = () => {
     history.push(routeToRules);
   };
@@ -265,17 +247,13 @@ export const RuleDetails: React.FunctionComponent<RuleDetailsProps> = ({
   };
 
   const onEditRuleClick = () => {
-    if (!isUsingRuleCreateFlyout) {
-      navigateToApp('management', {
-        path: `insightsAndAlerting/triggersActions/${getEditRuleRoute(rule.id)}`,
-        state: {
-          returnApp: 'management',
-          returnPath: `insightsAndAlerting/triggersActions/${getRuleDetailsRoute(rule.id)}`,
-        },
-      });
-    } else {
-      setEditFlyoutVisibility(true);
-    }
+    navigateToApp('management', {
+      path: `insightsAndAlerting/triggersActions/${getEditRuleRoute(rule.id)}`,
+      state: {
+        returnApp: 'management',
+        returnPath: `insightsAndAlerting/triggersActions/${getRuleDetailsRoute(rule.id)}`,
+      },
+    });
   };
 
   const editButton = hasEditButton ? (
@@ -292,19 +270,6 @@ export const RuleDetails: React.FunctionComponent<RuleDetailsProps> = ({
           defaultMessage="Edit"
         />
       </EuiButtonEmpty>
-      {editFlyoutVisible && (
-        <RuleEdit
-          initialRule={rule}
-          onClose={() => {
-            setInitialRule(rule);
-            setEditFlyoutVisibility(false);
-          }}
-          actionTypeRegistry={actionTypeRegistry}
-          ruleTypeRegistry={ruleTypeRegistry}
-          ruleType={ruleType}
-          onSave={setRule}
-        />
-      )}
     </>
   ) : null;
