--- x-pack/platform/plugins/shared/triggers_actions_ui/public/application/sections/rule_form/get_initial_interval.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { parseDuration } from '@kbn/alerting-plugin/common';
-import { DEFAULT_RULE_INTERVAL } from '../../constants';
-
-export function getInitialInterval(minimumScheduleInterval?: string) {
-  if (minimumScheduleInterval) {
-    // return minimum schedule interval if it is larger than the default
-    if (parseDuration(minimumScheduleInterval) > parseDuration(DEFAULT_RULE_INTERVAL)) {
-      return minimumScheduleInterval;
-    }
-  }
-  return DEFAULT_RULE_INTERVAL;
-}
