--- x-pack/platform/plugins/shared/triggers_actions_ui/public/application/sections/rule_form/rule_form.scss
@@ -1,4 +1,0 @@
-.triggersActionsUI__ruleTypeNodeHeading {
-  margin-left: $euiSizeS;
-  margin-right: $euiSizeS;
-}
