+++ x-pack/platform/plugins/shared/triggers_actions_ui/public/application/sections/rules_list/components/rules_list.tsx
@@ -14,16 +14,7 @@ import { FormattedMessage } from '@kbn/i18n-react';
 import { toMountPoint } from '@kbn/react-kibana-mount';
 import { parseRuleCircuitBreakerErrorMessage } from '@kbn/alerting-plugin/common';
 import { RuleTypeModal } from '@kbn/response-ops-rule-form';
-import React, {
-  lazy,
-  useEffect,
-  useState,
-  ReactNode,
-  useCallback,
-  useMemo,
-  useRef,
-  Suspense,
-} from 'react';
+import React, { useEffect, useState, ReactNode, useCallback, useMemo, useRef } from 'react';
 import {
   EuiSpacer,
   EuiPageTemplate,
@@ -37,14 +28,12 @@ import { useHistory } from 'react-router-dom';
 
 import {
   RuleExecutionStatus,
-  ALERTING_FEATURE_ID,
   RuleExecutionStatusErrorReasons,
   RuleLastRunOutcomeValues,
 } from '@kbn/alerting-plugin/common';
 import {
   RuleCreationValidConsumer,
   ruleDetailsRoute as commonRuleDetailsRoute,
-  STACK_ALERTS_FEATURE_ID,
   getCreateRuleRoute,
   getEditRuleRoute,
 } from '@kbn/rule-data-utils';
@@ -111,11 +100,6 @@ import { RulesSettingsLink } from '../../../components/rules_setting/rules_setti
 import { useRulesListUiState as useUiState } from '../../../hooks/use_rules_list_ui_state';
 import { useRulesListFilterStore } from './hooks/use_rules_list_filter_store';
 
-// Directly lazy import the flyouts because the suspendedComponentWithProps component
-// cause a visual hitch due to the loading spinner
-const RuleAdd = lazy(() => import('../../rule_form/rule_add'));
-const RuleEdit = lazy(() => import('../../rule_form/rule_edit'));
-
 export interface RulesListProps {
   ruleTypeIds?: string[];
   consumers?: string[];
@@ -142,7 +126,6 @@ export interface RulesListProps {
   onRefresh?: (refresh: Date) => void;
   setHeaderActions?: (components?: React.ReactNode[]) => void;
   initialSelectedConsumer?: RuleCreationValidConsumer | null;
-  useNewRuleForm?: boolean;
 }
 
 export const percentileFields = {
@@ -184,8 +167,6 @@ export const RulesList = ({
   onTypeFilterChange,
   onRefresh,
   setHeaderActions,
-  initialSelectedConsumer = STACK_ALERTS_FEATURE_ID,
-  useNewRuleForm = false,
 }: RulesListProps) => {
   const history = useHistory();
   const kibanaServices = useKibana().services;
@@ -205,10 +186,6 @@ export const RulesList = ({
   const [inputText, setInputText] = useState<string>(searchFilter);
 
   const [ruleTypeModalVisible, setRuleTypeModalVisibility] = useState<boolean>(false);
-  const [ruleTypeIdToCreate, setRuleTypeIdToCreate] = useState<string | undefined>(undefined);
-  const [ruleFlyoutVisible, setRuleFlyoutVisibility] = useState<boolean>(false);
-  const [editFlyoutVisible, setEditFlyoutVisibility] = useState<boolean>(false);
-  const [currentRuleToEdit, setCurrentRuleToEdit] = useState<RuleTableItem | null>(null);
   const [itemIdToExpandedRowMap, setItemIdToExpandedRowMap] = useState<Record<string, ReactNode>>(
     {}
   );
@@ -216,7 +193,6 @@ export const RulesList = ({
   const cloneRuleId = useRef<null | string>(null);
 
   const isRuleStatusFilterEnabled = getIsExperimentalFeatureEnabled('ruleStatusFilter');
-  const isUsingRuleCreateFlyout = getIsExperimentalFeatureEnabled('isUsingRuleCreateFlyout');
 
   const [percentileOptions, setPercentileOptions] =
     useState<EuiSelectableOption[]>(initialPercentileOptions);
@@ -321,18 +297,13 @@ export const RulesList = ({
   });
 
   const onRuleEdit = (ruleItem: RuleTableItem) => {
-    if (!isUsingRuleCreateFlyout && useNewRuleForm) {
-      navigateToApp('management', {
-        path: `insightsAndAlerting/triggersActions/${getEditRuleRoute(ruleItem.id)}`,
-        state: {
-          returnApp: 'management',
-          returnPath: `insightsAndAlerting/triggersActions/rules`,
-        },
-      });
-    } else {
-      setEditFlyoutVisibility(true);
-      setCurrentRuleToEdit(ruleItem);
-    }
+    navigateToApp('management', {
+      path: `insightsAndAlerting/triggersActions/${getEditRuleRoute(ruleItem.id)}`,
+      state: {
+        returnApp: 'management',
+        returnPath: `insightsAndAlerting/triggersActions/rules`,
+      },
+    });
   };
 
   const onRunRule = async (id: string) => {
@@ -1029,54 +1000,15 @@ export const RulesList = ({
           <RuleTypeModal
             onClose={() => setRuleTypeModalVisibility(false)}
             onSelectRuleType={(ruleTypeId) => {
-              if (!isUsingRuleCreateFlyout) {
-                navigateToApp('management', {
-                  path: `insightsAndAlerting/triggersActions/${getCreateRuleRoute(ruleTypeId)}`,
-                });
-              } else {
-                setRuleTypeIdToCreate(ruleTypeId);
-                setRuleTypeModalVisibility(false);
-                setRuleFlyoutVisibility(true);
-              }
+              navigateToApp('management', {
+                path: `insightsAndAlerting/triggersActions/${getCreateRuleRoute(ruleTypeId)}`,
+              });
             }}
             http={http}
             toasts={toasts}
             registeredRuleTypes={ruleTypeRegistry.list()}
             filteredRuleTypes={filteredRuleTypes}
           />
-        )}
-        {ruleFlyoutVisible && (
-          <Suspense fallback={<div />}>
-            <RuleAdd
-              consumer={ALERTING_FEATURE_ID}
-              onClose={() => {
-                setRuleFlyoutVisibility(false);
-              }}
-              actionTypeRegistry={actionTypeRegistry}
-              ruleTypeRegistry={ruleTypeRegistry}
-              ruleTypeIndex={ruleTypesState.data}
-              onSave={refreshRules}
-              initialSelectedConsumer={initialSelectedConsumer}
-              ruleTypeId={ruleTypeIdToCreate}
-              canChangeTrigger={false}
-            />
-          </Suspense>
-        )}
-        {editFlyoutVisible && currentRuleToEdit && (
-          <Suspense fallback={<div />}>
-            <RuleEdit
-              initialRule={currentRuleToEdit}
-              onClose={() => {
-                setEditFlyoutVisibility(false);
-              }}
-              actionTypeRegistry={actionTypeRegistry}
-              ruleTypeRegistry={ruleTypeRegistry}
-              ruleType={
-                ruleTypesState.data.get(currentRuleToEdit.ruleTypeId) as RuleType<string, string>
-              }
-              onSave={refreshRules}
-            />
-          </Suspense>
         )}
       </EuiPageTemplate.Section>
     </>
