+++ x-pack/platform/plugins/shared/triggers_actions_ui/public/types.ts
@@ -398,7 +398,6 @@ export interface RuleDefinitionProps<Params extends RuleTypeParams = RuleTypePar
   onEditRule: () => Promise<void>;
   hideEditButton?: boolean;
   filteredRuleTypes?: string[];
-  useNewRuleForm?: boolean;
 }
 
 export enum Percentiles {
