+++ x-pack/test_serverless/functional/config.base.ts
@@ -38,11 +38,6 @@ export function createTestConfig<TServices extends {} = typeof services>(
         serverArgs: [
           ...svlSharedConfig.get('kbnTestServer.serverArgs'),
           `--serverless=${options.serverlessProject}`,
-          // Ensures the existing E2E tests are backwards compatible with the old rule create flyout
-          // Remove this experiment once all of the migration has been completed
-          `--xpack.trigger_actions_ui.enableExperimental=${JSON.stringify([
-            'isUsingRuleCreateFlyout',
-          ])}`,
           ...(options.kbnServerArgs ?? []),
         ],
       },
